refactor: move wordle into its own folder moved wordle related components into its own folder, and extracted wordle api into it's own router
