+++ client/src/pages/Home.js
@@ -6,12 +6,6 @@ function Home() {
     return (
         <div>
             <h1>Project Home!</h1>
-            {/** Link to List.js */}
-            <Link to={'./list'}>
-                <button>
-                    My List
-                </button>
-            </Link>
             <Link to={'./wordle'}>
                 <button>
                     Wordle
